pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "pypdf>=6.7.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
]
